+++ CHANGELOG.md
@@ -35,6 +35,10 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 # Changelog
 
+## Unreleased
+
+- Implement with_market_incentives query param for `/pools` that formats fees and APRs data on pools.
+
 ## v25.8.1
 
 - Bump sqsdomain